Making CustomGroupingTest extend an abstract superclass
